Move AllObjects code to separate file, order objects per type
